Fix PR feedback: rename prompt_primitives → primitive_breakdown, append results, update README instructions
